move the database file (BufferScroll.bin.gz) to the Settings directory
The database file stores saved settings per buffer and is very
similar in nature to Sublime Text's own session mechanism. Storing
it in the same directory as the ST2 session is useful for people who
store their User package in Dropbox, since it lies outside of the
Packages directory and does not need to be synchronized between
machines.